Modifikacija koda

+++ src/main/java/spring/projekat/controller/DokumentController.java
@@ -1,40 +1,56 @@
 package spring.projekat.controller;
 
-import java.util.List;
+
+import java.util.HashSet;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.annotations.ApiOperation;
 import spring.projekat.model.Dokument;
+import spring.projekat.model.Roba;
+import spring.projekat.model.Stavka_Dokumenta;
 import spring.projekat.service.DokumentService;
+import spring.projekat.service.RobaService;
+import spring.projekat.service.Stavka_DokumentaService;
 
 @RestController
+@RequestMapping("/api/")
 public class DokumentController {
 
 	
 	@Autowired
 	DokumentService dokumentService;	
 	
+	@Autowired
+	Stavka_DokumentaService sds;
+	
+	@Autowired
+	RobaService robaService;
+	
 	//GET/ALL
 	@ApiOperation(value = "Pronalazi sva dokumenta")
-	@RequestMapping(value = "api/dokuments", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_VALUE})
-	public ResponseEntity<List<Dokument>> getAllDokuments(){
-		List<Dokument> dokuments = dokumentService.findAll();
+	@GetMapping(value = "dokuments", produces = {MediaType.APPLICATION_JSON_VALUE})
+	public ResponseEntity<Iterable<Dokument>> getAllDokuments(){
+		Iterable<Dokument> dokuments = dokumentService.findAll();
 		
 		return new ResponseEntity<>(dokuments, HttpStatus.OK);
 	}
 	
 	//GET/ONE
 	@ApiOperation(value = "Pronalazi dokumenta po id-u", notes = "Pomocu id-a pronalazi odredjeni dokument")
-	@RequestMapping(value = "api/dokument/{id}", method = RequestMethod.GET, produces = {MediaType.APPLICATION_JSON_VALUE})
+	@GetMapping(value = "dokument/{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
 	public ResponseEntity<Dokument> getDokument(@PathVariable Long id){
 		Dokument dokument = dokumentService.findOne(id);
 		
@@ -42,16 +58,17 @@ public class DokumentController {
 	}
 	
 	//POST
-	@RequestMapping(value = "api/dokuments", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Dokument> create(@RequestBody Dokument dokument) {
+	@PostMapping(value = "dokuments", consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<Dokument> create(@RequestBody Dokument dokument) {			
+		
 		Dokument retVal = dokumentService.save(dokument);
-
+		
 		return new ResponseEntity<>(retVal, HttpStatus.CREATED);
 	}
 	
 	
 	//PUT
-	@RequestMapping(value = "api/dokuments/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
+	@PutMapping(value = "dokument/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Dokument> update(@PathVariable Long id,
 			@RequestBody Dokument dokument) {
 		dokument.setId(id);
@@ -61,12 +78,16 @@ public class DokumentController {
 	}
 	
 	//DELETE
-	@RequestMapping(value = "api/dokuments/{id}", method = RequestMethod.DELETE)
+	@DeleteMapping(value = "dokument/{id}")
 	public ResponseEntity<Void> delete(@PathVariable Long id) {
 		Dokument dokument = dokumentService.findOne(id);
 		if (dokument != null) {
-			dokumentService.remove(id);
-			return new ResponseEntity<>(HttpStatus.OK);
+			if(dokument.getStavkaDokumenta().isEmpty()) {
+				dokumentService.remove(id);
+				return new ResponseEntity<>(HttpStatus.OK);
+			}else {
+				return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+			}
 		} else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
